Add files via upload
1. Write a static method isStrictlyIncreasing(double[] in) that returns true if each value in the given array is greater than
the value before it, or false otherwise.
2. Write a static method removeDuplicates(Character[] in) that returns a new array of the characters in the given array,
but without any duplicate characters. Always keep the first copy of the character and remove subsequent ones. For
example, if in contains b, d, a, b, f, a, g, a, a, and f, the method will return an array containing b, d a, f, and g. (Hint:
One way to solve this problem is to create a boolean array of the same size as the given array in and use it to keep
track of which characters to keep. The values in the new boolean array will determine the size of the array to return.)
3. Write a static method remove(intv,int[] in) that will return a new array of the integers in the given array, but with the
value v removed. For example, if v is 3 and in contains 0, 1, 3, 2, 3, 0, 3, and 1, the method will return an array
containing 0, 1, 2, 0, and 1.
4. Suppose that we are selling boxes of candy for a fund-raiser. We have five kinds of candy to sell: Mints, Chocolates
with Nuts, Chewy Chocolates, Dark Chocolate Creams, and Sugar-Free Suckers. We will record a customer’s order as
an array of five integers, representing the number of boxes of each kind of candy. Write a static method
combineOrder that takes two orders as its arguments and returns an array that represents the combined orders. For
example, if order1 contains 0, 0, 3, 4, and 7, and order2 contains 0, 4, 0, 1, and 2, the method should return an array
containing 0, 4, 3, 5, and 9.