Add files via upload
1. Create an abstract class PayCalculator that has an attribute payRate given in dollars per hour. The
class should also have a method computePay(hours) that returns the pay for a given amount of
time.
2. Derive a class RegularPay from PayCalculator, as described in the previous exercise. It should have
a constructor that has a parameter for the pay rate. It should not override any of the methods. Then
derive a class HazardPay from PayCalculator that overrides the computePay method. The new
method should return the amount returned by the base class method multiplied by 1.5.
3. Create an abstract class DiscountPolicy. It should have a single abstract method computeDiscount
that will return the discount for the purchase of a given number of a single item. The method has two
parameters, count and itemCost.
4. Derive a class BulkDiscount from DiscountPolicy, as described in the previous exercise. It should
have a constructor that has two parameters, minimum and percent. It should define the method
computeDiscount so that if the quantity purchased of an item is more than minimum, the discount is
percent percent.
5. Create an interface MessageEncoder that has a single abstract method encode(plainText), where
plainText is the message to be encoded. The method will return the encoded message.
6. Create a class SubstitutionCipher that implements the interface MessageEncoder, as described in the
previous exercise. The constructor should have one parameter called shift. Define the method encode so
that each letter is shifted by the value in shift. For example, if shift is 3, a will be replaced by d, b will be
replaced by e, c will be replaced by f, and so on. Hint: You may wish to define a private method that shifts
a single character.